+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.7.3]
+
+### Added
+- New tile/touch action mushymato.MMAP_WarpBuilding [X Y], warp that will put player into the building on this tile, for use with buildings.
+- New tile/touch action mushymato.MMAP_MagicWarpBuilding [X Y], magic warp that will put player into the building on this tile, for use with buildings.
+
+## [1.7.2]
+
+### Fixed
+- NRE on map reload.
+
 ## [1.7.1]
 
 ### Changed